refactor: use package-level config imports
