+++ core/app_factory/__init__.py
@@ -97,7 +97,7 @@ from flask_caching import Cache
 
 from components.ui.navbar import create_navbar_layout
 from config.complete_service_registration import register_all_application_services
-from config.config import get_config
+from config import get_config
 from core.callback_registry import GlobalCallbackRegistry, _callback_registry
 from core.performance_monitor import DIPerformanceMonitor
 from core.service_container import ServiceContainer
