+++ core/plugins/unified_registry.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING, Any, List, Optional
 
 from dash import Dash
 
-from config.config import ConfigManager
+from config import ConfigManager
 from core.callback_manager import CallbackManager
 from core.service_container import ServiceContainer
 from core.plugins.manager import ThreadSafePluginManager
