+++ tests/test_plugin_manager.py
@@ -2,7 +2,7 @@ import sys
 import time
 from pathlib import Path
 
-from config.config import create_config_manager
+from config import create_config_manager
 from core.service_container import ServiceContainer
 from core.plugins.manager import ThreadSafePluginManager as PluginManager
 from services.data_processing.core.protocols import PluginMetadata
