+++ tests/test_plugin_manager_core.py
@@ -3,7 +3,7 @@ import sys
 import types
 from pathlib import Path
 
-from config.config import create_config_manager
+from config import create_config_manager
 from core.service_container import ServiceContainer
 from core.plugins.manager import ThreadSafePluginManager as PluginManager
 from services.data_processing.core.protocols import PluginMetadata, PluginStatus
