+++ tests/test_plugin_manager_thread.py
@@ -1,6 +1,6 @@
 import time
 
-from config.config import create_config_manager
+from config import create_config_manager
 from core.service_container import ServiceContainer
 from core.plugins.manager import ThreadSafePluginManager as PluginManager
